The old change broken building any other RPi based project
this would be needed if its ever taken upstream :)

+++ packages/tools/bcm2835-bootloader/package.mk
@@ -45,7 +45,10 @@ makeinstall_target() {
     cp -PRv bootcode.bin $INSTALL/usr/share/bootloader
     cp -PRv fixup_x.dat $INSTALL/usr/share/bootloader/fixup.dat
     cp -PRv start_x.elf $INSTALL/usr/share/bootloader/start.elf
-    cp -PRv dt-blob.bin $INSTALL/usr/share/bootloader/dt-blob.bin
+    
+    if [ -f $PROJECT_DIR/$PROJECT/bootloader/dt_blob.dts ]; then
+        cp -PRv dt-blob.bin $INSTALL/usr/share/bootloader/dt-blob.bin
+    fi
 
     cp -PRv $PKG_DIR/scripts/update.sh $INSTALL/usr/share/bootloader
     cp -PRv $PKG_DIR/files/3rdparty/bootloader/config.txt $INSTALL/usr/share/bootloader
